cleaned up devices interfaces to use common namespace like the other projects
