+++ Vapour.Server.Host/Controller/ControllerService.cs
@@ -1,6 +1,6 @@
 ﻿using Vapour.Server.Controller;
 using Vapour.Shared.Devices.HostedServices;
-using Vapour.Shared.Devices.Interfaces.Services;
+using Vapour.Shared.Devices.Services;
 
 namespace Vapour.Server.Host.Controller;
