+++ Vapour.Server.Interfaces/Controller/ControllerConnectedMessage.cs
@@ -1,4 +1,4 @@
-﻿using Vapour.Shared.Devices.Interfaces.HID;
+﻿using Vapour.Shared.Devices.HID;
 
 namespace Vapour.Server.Controller;
