+++ Vapour.Shared.Devices.Interfaces/HID/CompatibleHidDeviceInputReport.cs
@@ -1,4 +1,4 @@
-﻿namespace Vapour.Shared.Devices.Interfaces.HID;
+﻿namespace Vapour.Shared.Devices.HID;
 
 public enum DPadDirection
 {
