+++ Vapour.Shared.Devices.Interfaces/HID/DeviceNotificationListenerSubscriber.cs
@@ -1,4 +1,4 @@
-﻿namespace Vapour.Shared.Devices.Interfaces.HID;
+﻿namespace Vapour.Shared.Devices.HID;
 
 public interface IDeviceNotificationListenerSubscriber
 {
