+++ Vapour.Shared.Devices.Interfaces/HID/ICompatibleHidDevice.cs
@@ -1,6 +1,6 @@
 ﻿using System.Net.NetworkInformation;
 
-namespace Vapour.Shared.Devices.Interfaces.HID;
+namespace Vapour.Shared.Devices.HID;
 
 public interface ICompatibleHidDevice : IDisposable
 {
