+++ Vapour.Shared.Devices.Interfaces/HID/InputDeviceType.cs
@@ -1,4 +1,4 @@
-﻿namespace Vapour.Shared.Devices.Interfaces.HID;
+﻿namespace Vapour.Shared.Devices.HID;
 
 public enum InputDeviceType : uint
 {
