+++ Vapour.Shared.Devices/HID/Devices/DualSenseCompatibleHidDevice.cs
@@ -1,7 +1,6 @@
 ﻿using Microsoft.Extensions.Logging;
 
 using Vapour.Shared.Devices.HID.Devices.Reports;
-using Vapour.Shared.Devices.Interfaces.HID;
 
 namespace Vapour.Shared.Devices.HID.Devices;
