+++ Vapour.Shared.Devices/HID/Devices/JoyConCompatibleHidDevice.cs
@@ -1,7 +1,6 @@
 ﻿using System.Net.NetworkInformation;
 
 using Vapour.Shared.Devices.HID.Devices.Reports;
-using Vapour.Shared.Devices.Interfaces.HID;
 
 namespace Vapour.Shared.Devices.HID.Devices;
